Move common code to PixelClusterizerBase
