+++ RecoLocalTracker/SiPixelClusterizer/plugins/PixelClusterizerBase.h
@@ -1,13 +1,14 @@
 #ifndef RecoLocalTracker_SiPixelClusterizer_PixelClusterizerBase_H
 #define RecoLocalTracker_SiPixelClusterizer_PixelClusterizerBase_H
 
+#include <vector>
+
+#include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationServiceBase.h"
 #include "DataFormats/Common/interface/DetSetVector.h"
 #include "DataFormats/Common/interface/DetSetVectorNew.h"
 #include "DataFormats/SiPixelCluster/interface/SiPixelCluster.h"
 #include "DataFormats/SiPixelDigi/interface/PixelDigi.h"
 #include "DataFormats/TrackerCommon/interface/TrackerTopology.h"
-#include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationServiceBase.h"
-#include <vector>
 
 class PixelGeomDetUnit;
 
@@ -20,29 +21,38 @@ public:
   typedef edmNew::DetSet<SiPixelCluster>::const_iterator ClusterIterator;
 
   struct AccretionCluster {
-    typedef unsigned short UShort;
-    static constexpr UShort MAXSIZE = 256;
-    UShort adc[MAXSIZE];
-    UShort x[MAXSIZE];
-    UShort y[MAXSIZE];
-    UShort xmin = 16000;
-    UShort ymin = 16000;
+    static constexpr uint16_t MAXSIZE = 256;
+    uint16_t adc[MAXSIZE];
+    uint16_t x[MAXSIZE];
+    uint16_t y[MAXSIZE];
+    uint16_t xmin = 16000;
+    uint16_t ymin = 16000;
     unsigned int isize = 0;
-    unsigned int curr = 0;
+    int charge = 0;
 
     // stack interface (unsafe ok for use below)
-    UShort top() const { return curr; }
+    unsigned int curr = 0;
+    uint16_t top() const { return curr; }
     void pop() { ++curr; }
     bool empty() { return curr == isize; }
 
-    bool add(SiPixelCluster::PixelPos const& p, UShort const iadc) {
+    void clear() {
+      xmin = 16000;
+      ymin = 16000;
+      isize = 0;
+      charge = 0;
+      curr = 0;
+    }
+
+    bool add(SiPixelCluster::PixelPos const& p, uint16_t const iadc) {
       if (isize == MAXSIZE)
         return false;
-      xmin = std::min(xmin, (unsigned short)(p.row()));
-      ymin = std::min(ymin, (unsigned short)(p.col()));
+      xmin = std::min<uint16_t>(xmin, p.row());
+      ymin = std::min<uint16_t>(ymin, p.col());
       adc[isize] = iadc;
       x[isize] = p.row();
       y[isize++] = p.col();
+      charge += iadc;
       return true;
     }
   };
